requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.3
    # via web3
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
bitarray==2.9.2
    # via eth-account
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
cytoolz==0.12.3
    # via eth-utils
environs==10.3.0
    # via
    #   -r requirements.in
    #   eth-prototype
eth-abi==5.0.0
    # via
    #   eth-account
    #   web3
eth-account==0.11.0
    # via web3
eth-hash[pycryptodome]==0.6.0
    # via
    #   eth-hash
    #   eth-utils
    #   web3
eth-keyfile==0.7.0
    # via eth-account
eth-keys==0.5.0
    # via
    #   eth-account
    #   eth-keyfile
eth-prototype[web3]==1.0.0
    # via
    #   -r requirements.in
    #   eth-prototype
eth-rlp==1.0.1
    # via eth-account
eth-typing==4.0.0
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==3.0.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
future==0.18.3
    # via m9g
future-fstrings==1.2.0
    # via m9g
hexbytes==0.3.1
    # via
    #   eth-account
    #   eth-prototype
    #   eth-rlp
    #   web3
idna==3.6
    # via
    #   requests
    #   yarl
jsonschema==4.21.1
    # via web3
jsonschema-specifications==2023.12.1
    # via jsonschema
lru-dict==1.2.0
    # via web3
m9g==0.0.6
    # via
    #   -r requirements.in
    #   eth-prototype
marshmallow==3.20.2
    # via environs
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
packaging==23.2
    # via marshmallow
parsimonious==0.9.0
    # via eth-abi
protobuf==4.25.3
    # via web3
pycryptodome==3.20.0
    # via
    #   eth-hash
    #   eth-keyfile
python-dotenv==1.0.1
    # via environs
pyunormalize==15.1.0
    # via web3
pyyaml==6.0.1
    # via -r requirements.in
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.12.25
    # via parsimonious
requests==2.31.0
    # via
    #   eth-prototype
    #   web3
rlp==4.0.0
    # via
    #   eth-account
    #   eth-rlp
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
toolz==0.12.1
    # via cytoolz
typing-extensions==4.9.0
    # via
    #   eth-rlp
    #   web3
urllib3==2.2.1
    # via requests
web3==6.15.1
    # via eth-prototype
websockets==12.0
    # via web3
yarl==1.9.4
    # via aiohttp
